Update Reserve Margin Constraint
Several updates are made to the reserve margin constraint:

1. Exhange techs (i.e. transmission) are now accounted for.

Prior to this commit, the reserve margin constraint for region R1
was defined over all technologies in R1 belonging to the tech_reserve
set. Exchange techs, e.g. electricity transmission, however, are indexed
over region pairs, e.g. R1-R2, and are therefore not included in the
current constraint.

This commit exptends the constraint to also include exchange techs.

First, the capacity credit of exchange techs are explicitly accounted for
in the constraint. For R1, this is done by searching through all technologies
in the tech_reserve set with region XX-R1, where XX is any other model
region. This uses the temoa convention that exchange techs belonging to region
R2-R1 import commodities into R1 and exchange techs belonging to region R1-R2
export commodities out of R1.

Second, the activity of exchange techs are explicitly accounted for in the
determination of the total generation in each time step. Electricity imports
are added to the total generation and electricity exports are subtracted from
the total generation. This gives a better picture of the net regional generation,
which temoa uses as a proxy for regional electricity demand.

2. Account for storage charging

Electricty flowing into a storage technology is subtracted from the
load in the peak load calculation.