Do not report arbitrarily small capacities in outputs
Due to finite precision, the solver will output very small
non-zero values for certain technologies' capacities (this is effectively
zero). Very small capacity values can cause errors in myopic formulations
because these will be passed onto subsequent runs as Existing Capacity in the
model. This then results in very poor conditioning of the LP matrix.

We prevent this by defining a new parameter, capacity_epsilon. Capacities below
which are not printed to the output file (and take on the value of 0).