Cleaned up tests folder
Tests running (except myopic, which is "on hold").  Added testing_outputs folder to catch database writes that occur during tests to preclude the input databases from being written to.  The output folder will automatically be populated with db copies if they aren't already there--and they should NOT be added to VCS