Basic Test of Myopic Mode Updated and Running (passing)
The full test for myopic mode now runs and re-uses the utopia database in the testing_outputs folder (not great, but a good placeholder for now).  Bug still exists in main myopic code that forces input and output db to be one-in-the-same.  Also (for now) the name of the config file for myopic usage must be exactly "config_sample" because internally constructed variants of that such as "config_sample_myopic" are tested for *by name* in several code locations.  No other names for myopic mode config files are functional (yet)