Removed the Reference for VintageDefaults for Cost
Vintage defaults for fixed and variable costs are no longer part of the db schema.  All periods must now be specified, so these auxiliary parameters which were used as defaults are not needed, along with the build action to create costs.  Updated the model set_consistency tests to reflect the removal of the index sets for these.  All tests pass.