Added New Scenario Mediumville to Testing Suite
Mediumville is intended to exercise ALL (valid) tables in the input database, which is extremely helpful in QA'ing all of the model sets, finding the defunct ones for weeding out, and ensuring that follow-on actions do not corrupt all model sets for a running model.  Mediumville solves, and all tests are passing.  The re-making of the .json sets for others was done along with this, and some re-ordering caused their update, but both this and previous pass, meaning the only delta for Utopia and TestSystem is the ordering in the json, which is inconsequential.