Further Updates (#65)
* Added Logging Warnings to MaxResource and ResourceBound constraints

Both of these constraints will now throw logging errors.  They both need attention after DB schema stabilizes

* Update to Screening of Exchange Techs

Added a condition to screening of ExchangeTechs in initialize file.  We only need to check periods *on/after* the introduction of the exchange tech for problems, not before.

* New Cost Table Prototype Included

Added a new cost table that captures all model-view costs and undiscounted costs

* New Cost Table Working

Added a new cost table that captures all model-view costs and undiscounted costs

* Added Tests for ExchangeTechCostLedger

The ledger is responsible for splitting costs in Exchange Techs.  Tests were added, bugs found and squashed.

* Got Duals and save_LP_files Working

The selections (in config) to save Duals and save the LP files are now working/tested for both PF and myopic.

* Test Folder Cleanup

removing old .dat files and other unused files.

* Updated testing DB's

Migrated them to include the new source commodity label and unlimited resource for compatibility

* Added Basic Test for Myopic Run

* Chops While working on Myopic Run

* Fixes to Handling of tables in hybrid_loader.py without myopic index.

Repaired previously introduced error.  Test passes.

* Some Minor Changes and addition of M.troubleshooting

Added ability to turn on "troubleshooting" capacity and flow issues within temoa_model.py.

* Some Minor Changes and addition of M.troubleshooting

Added ability to turn on "troubleshooting" capacity and flow issues within temoa_model.py.

* Myopic Mode Test Working

Update to tests to ensure myopic test works (in very basic format.)  Also changed db options for myopic to only work from 1 database and not copy it for the user (as it was attempting to do before.)  Added VACUUM to myopic sequence loop to compact DB periodically.

* Adjusted Undiscounted Investment Cost Calculation.

Changed the calculation of undiscounted to only capture payments within periods in the modeling horizon.

* Correction to Undiscounted Loan Cost Formulation

* Minor Housekeeping Edits

* Added ModelNetworkData Element

This simple structure can be built from a TemoaModel or from a DB connection to drive the source_check.py file.  Intent is to be able to drive it from DB and use that as end-product of the source checking as a filter for the hybrid_loader.py.

* Added Testing of Network Model Data

* Added Additional Test for Source Tracing from DB

* Split out the commodity_network.py File

* Updated Utopia Test DB to include Output_Cost_2 Table

This was needed because it was being added "on the fly" by myopic action, but it should be there as default.

* Reorganizing

Moved Myopic to Extensions folder.  Place sample config in "my configs" folder.

* Added config_sample.toml to my_configs folder and Partial Update to README.md